Fix flaky cache-storage-reporting.https.html
The test is flaky, because it rely on a 4s timeout. This is not enough
always enough.

This patch splits this test into 4. One for every execution context:
- Document
- ServiceWorker
- DedicatedWorker
- SharedWorker

This allows the timeout to become 'infinite', while still making sure
the passing tests are still executed. Currently, only the SharedWorker
hasn't been implemented yet. As a result it is marked as [ Timeout ]

Bug: 1061029
Change-Id: If9a8e7b65d91cccffeeedb3f1a13d5e6e32e41bf